bug: vertical sync happened 7 to 8x too fast.
Caused by hsyncgen.rastclken signal occupying an entire character cell.
Fix is to AND it with hsyncgen.xclken.

NOTE: There is still an additional bug in this version: the total adjust
setting **adds** to the width of the vertical sync, instead of (or
perhaps in addition to?) delaying reset of vertical counter to 0.