Vsync (#17).  Fixes #15
* Switch from count-up to count-down character counter.

The VDC appears to use down-counters for triggering events, such as
flagging when to start shifting dots to teh screen, even though some
registers (e.g., light-pen) appear to make use of up-counters.  I don't
anticipate using light-pens with the VDC-II (though the registers will
remain reserved for future restoration of the facility if someone wants
it).  Therefore, switching to using down-counters optimizes for the more
common case.

I think.

VDC documentation can be pretty confusing at times, and low-level
descriptions of the VDC seems to simply not exist.

* Yet another sync generator.

This one more closely models the origial 6545 CRTC behavior, which is
what the 8563 inherits.  If I can add support for horizontal total
adjustment, then I can reuse the same module for vertical sync as well.

* Horizontal Total Adjust support.

* dot clock enable input.

Added a master enable signal to control which dot clocks are relevant,
even when hclken output isn't asserted.  This is required because the
total adjust functionality always works on units of pixels/lines, not on
units of characters.

* Refactor HSYNC generator into something generic.

* Remove old sync generator logic

* refactor

* This should complete the VSYNC interface.

In addition to generating both HSYNC and VSYNC, this circuit should also
drive the display with a test pattern which illustrates where character
cells reside on the playfield.

* Bug fix: only low bit of register address was significant.

* bug: vertical sync happened 7 to 8x too fast.

Caused by hsyncgen.rastclken signal occupying an entire character cell.
Fix is to AND it with hsyncgen.xclken.

NOTE: There is still an additional bug in this version: the total adjust
setting **adds** to the width of the vertical sync, instead of (or
perhaps in addition to?) delaying reset of vertical counter to 0.

* Fix vsync pulse width bug

* Restore grid patteron on test display