MPE (#19)
* First code for MPE

* Record prefetched data in holding register

* Let a host CPU read data from video memory.

* Implement update location registers

* Generate read command pulse from host interface

* Strobe to detect data port reads

* Prevent register contention for update location.

It's possible that a host processor will attempt to write a new pointer
value into R18/R19 while the VDC is trying to increment it.  Not
*entirely* arbitrarily, I decided the right path is to increment the
pointer.  This is because the VDC might be in the process of performing
a block write or block copy operation while R18/R19 are written to.  If
the write were to succeed instead, the block operation is guaranteed to
corrupt video memory.

* Support CPU write to CPU data register.

* Record data written to R31 for the MPE.

* Adhere to Wishbone naming conventions

* Complete block write/copy engine

* Expose byte counter and decrement control for MPE.

* Bugfix: forgot to decrement byte counter

* First attempts to integrate MPE and 16KB of video RAM.

1. Route VSYNC flag to bit 6 of status register.
2. Route READY flag to bit 7 of status register.
3. Introduce MPE.
4. Introduce block RAM direct-coupled to MPE.  No arbiter yet.

* Forgot to include ram.py