Implemented ROM-based "strip buffer".
The ROM-based strip buffer is designed to test our ability to put pixels
on the display from "video memory" without implementing the complexity
of a memory arbiter and video fetch engine just yet.

It confirms that attributes are properly aligned with pixels, and that
horizontal smooth scroll continues to work as expected in the face of
these new changes.