Bitmap mode (#27)
* Record an easy to use blank module template

* remove phantom I/Os

* WIP: address generator logic for video fetch.

I don't do anything with the fetched data yet; but this *seems* like it
should work to fetch color or monochrome text or bitmapped mode data.

* Data receiver logic

* execute permission on tester.sh

* Initial work on dot shifter.

This implements the core logic for horizontal smooth scroll.  This logic
will be instrumental in building out the remainder of the dot shifter.

* Implemented ROM-based "strip buffer".

The ROM-based strip buffer is designed to test our ability to put pixels
on the display from "video memory" without implementing the complexity
of a memory arbiter and video fetch engine just yet.

It confirms that attributes are properly aligned with pixels, and that
horizontal smooth scroll continues to work as expected in the face of
these new changes.

* Block RAM arbiter

* Refactor shifter

* Pass asize through

* Refactor video fetch engine

* Strip buffer

* Expose atrbase to other subcircuits

* Reset atrptr at start of video frame

* Vertical smooth scroll, attr/char ptrs

* WIP: monochrome bitmap almost working.

Bitmap generation and vertical scrolling doesn't quite work.
If I enable attributes, I can see video memory does contain alternating
sequences of $AA and $55 bytes (see GREY.BAS to change these defaults).
However, instead of the stipple pattern I expected, I get one bitmap row
repeated for every raster row in a character row.  Except, that is, for
bitmap row 0.

Perhaps this has something to do with how chrptr is managed (in
shifter.py).

* Fixes bad fetch pointer values at top raster row.

Works for bitmap and character modes as far as I can tell so far.

* Record my test programs